+++ apps/docs/src/app/core/component-docs/dialog/dialog-docs.module.ts
@@ -17,6 +17,8 @@ import { FormModule } from '@fundamental-ngx/core/form';
 import { ListModule } from '@fundamental-ngx/core/list';
 import { DialogModule, DialogService } from '@fundamental-ngx/core/dialog';
 import { InputGroupModule } from '@fundamental-ngx/core/input-group';
+import { MultiInputModule } from '@fundamental-ngx/core';
+import { DialogInnerPopoverComponent, DialogInnerPopoverExampleComponent } from './examples/dialog-inner-popover/dialog-inner-popover.component';
 
 const routes: Routes = [
     {
@@ -37,19 +39,23 @@ const routes: Routes = [
         ListModule,
         DialogModule,
         InputGroupModule,
-        A11yModule
+        A11yModule,
+        MultiInputModule
     ],
     exports: [RouterModule],
     declarations: [
         examples,
         DialogDocsComponent,
-        DialogDocsHeaderComponent
+        DialogDocsHeaderComponent,
+        DialogInnerPopoverComponent,
+        DialogInnerPopoverExampleComponent
     ],
     entryComponents: [
         DialogExampleComponent,
         FirstDialogExampleComponent,
         SecondDialogExampleComponent,
-        DialogStackedExampleComponent
+        DialogStackedExampleComponent,
+        DialogInnerPopoverExampleComponent
     ],
     providers: [
         DialogService